speech-generator/pyproject.toml
[project]
name = "speech-generator"
version = "0.1.0"
description = ""
authors = [
    {name = "Ivan Babiankou",email = "ivan.babiankou@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10, <3.13"
dependencies = [
    "fastapi (>=0.121.3,<0.122.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "kokoro (>=0.9.4,<0.10.0)",
    "soundfile (>=0.13.1,<0.14.0)",
    "pika (>=1.3.2,<2.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
packages = [{include = "api"}]
